requirements.txt
Django
djangorestframework
drf-spectacular
django-filter
Pillow
django-cors-headers
